# swagger.yaml
swagger: "2.0"
info:
  description: "Bot framkework API specification"
  version: "1.0.0"
  title: "AMiner Bot Framework"
  contact:
    email: "ypr15@mails.tsinghua.edu.cn"
host: "api.example.com"
basePath: "/v1"
...
schemes:
- "https"
- "http"
paths:
...
  /agent/{agent_id}/entity:
    get:
      tags:
      - "entity"
      summary: "Returns all entity types"
      description: "Returns all entity types of this agent"
      operationId: "getEntityTypes"
      produces:
      - "application/json"
      parameters:
      - name: "agent_id"
        in: "path"
        description: "Agent id"
        required: true
        type: "string"
        
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EntityType"
            
      security:
      - api_key: []
    post:
      tags:
      - "entity"
      summary: "Create new entity types"
      description: "Create a new entity type for an agent"
      operationId: "createNewEntity"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "agent_id"
        in: "path"
        description: "Agent id"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Entity object that needs to be updated."
        required: true
        schema:
          type: "object"
          properties:
            name:
              type: "string"
            description:
              type: "string"
            entries:
              type: "array"
              items:
                type: "string"
        
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EntityType"
            
      security:
      - api_key: []
  /agent/{agent_id}/entity/{entity_id}:
    get:
      tags:
      - "entity"
      summary: "Return the EntityType with given id"
      description: "Return the EntityType with given id"
      operationId: "getEntityType"
      produces:
      - "application/json"
      parameters:
      - name: "agent_id"
        in: "path"
        description: "Agent id"
        required: true
        type: "string"
      - name: "entity_id"
        in: "path"
        description: "EntityType id"
        required: true
        type: "string"
        
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EntityType"
        400:
          description: "Client error"
          schema:
            $ref: '#/definitions/ApiErrorResponse'
            
      security:
      - api_key: []
    put:
      tags:
      - "entity"
      summary: "Update the EntityType with given id"
      description: "Update the EntityType with given id, no field is required."
      operationId: "updateEntityType"
      produces:
      - "application/json"
      parameters:
      - name: "agent_id"
        in: "path"
        description: "Agent id"
        required: true
        type: "string"
      - name: "entity_id"
        in: "path"
        description: "EntityType id"
        required: true
        type: "string"
      - name: "body"
        in: "body"
        description: "EntityType object to be updated"
        required: true
        schema:
          $ref: '#/definitions/EntityType'
        
      responses:
        200:
          description: "successful operation"
        400:
          description: "Client error"
          schema:
            $ref: '#/definitions/ApiErrorResponse'            
      security:
      - api_key: []
    delete:
      tags:
      - "entity"
      summary: "Deletes the EntityType with given id"
      description: "Deletes the EntityType with given id"
      operationId: "deleteEntityType"
      produces:
      - "application/json"
      parameters:
      - name: "agent_id"
        in: "path"
        description: "Agent id"
        required: true
        type: "string"
      - name: "entity_id"
        in: "path"
        description: "EntityType id"
        required: true
        type: "string"
        
      responses:
        200:
          description: "successful operation"
        400:
          description: "Client error"
          schema:
            $ref: '#/definitions/ApiErrorResponse'            
      security:
      - api_key: []
  /agent/{agent_id}/entity/{entity_id}/addEntries:
    post:
      tags:
      - "entity"
      summary: "Add entries to the EntityType"
      description: "Add entries to the EntityType"
      operationId: "addEntry"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "agent_id"
        in: "path"
        description: "Agent id"
        required: true
        type: "string"
      - name: "entity_id"
        in: "path"
        description: "EntityType id"
        required: true
        type: "string"
      - name: "body"
        in: "body"
        description: "Entries"
        required: true
        schema:
          type: "array"
          items:
            type: "string"
        
      responses:
        200:
          description: "successful operation"

        400:
          description: "Client error"
          schema:
            $ref: '#/definitions/ApiErrorResponse'
            
      security:
      - api_key: []
  /agent/{agent_id}/entity/{entity_id}/deleteEntries:
    post:
      tags:
      - "entity"
      summary: "Delete entries from the EntityType"
      description: "Delete entries from the EntityType"
      operationId: "deleteEntry"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "agent_id"
        in: "path"
        description: "Agent id"
        required: true
        type: "string"
      - name: "entity_id"
        in: "path"
        description: "EntityType id"
        required: true
        type: "string"
      - name: "body"
        in: "body"
        description: "Entries"
        required: true
        schema:
          type: "array"
          items:
            type: "string"
        
      responses:
        200:
          description: "successful operation"

        400:
          description: "Client error"
          schema:
            $ref: '#/definitions/ApiErrorResponse'
            
      security:
      - api_key: []
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "query"
definitions:
...
  EntityType:
    type: "object"
    properties:
      agent_id:
        type: "string"
      name:
        type: "string"
      description:
        type: "string"
      entries:
        type: "array"
        items:
          type: "string"
...
  ApiErrorResponse:
    type: "object"
    properties:
      type:
        type: "string"
      message:
        type: "string"
...